ida-exercise-week-2-student-repo-public-student/x/cosmosacademyexercisesimple/client/cli/query_world_params_2.go
package cli

import (
	"context"
	"strconv"

	"github.com/b9lab/cosmos-academy-exercise-simple/x/cosmosacademyexercisesimple/types"
	"github.com/cosmos/cosmos-sdk/client"
	"github.com/cosmos/cosmos-sdk/client/flags"
	"github.com/spf13/cobra"
)

func CmdListWorldParams2() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "list-world-params-2",
		Short: "list all WorldParams2",
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			pageReq, err := client.ReadPageRequest(cmd.Flags())
			if err != nil {
				return err
			}

			queryClient := types.NewQueryClient(clientCtx)

			params := &types.QueryAllWorldParams2Request{
				Pagination: pageReq,
			}

			res, err := queryClient.WorldParams2All(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddPaginationFlagsToCmd(cmd, cmd.Use)
	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}

func CmdShowWorldParams2() *cobra.Command {
	cmd := &cobra.Command{
		Use:   "show-world-params-2 [id]",
		Short: "shows a WorldParams2",
		Args:  cobra.ExactArgs(1),
		RunE: func(cmd *cobra.Command, args []string) error {
			clientCtx := client.GetClientContextFromCmd(cmd)

			queryClient := types.NewQueryClient(clientCtx)

			id, err := strconv.ParseUint(args[0], 10, 64)
			if err != nil {
				return err
			}

			params := &types.QueryGetWorldParams2Request{
				Id: id,
			}

			res, err := queryClient.WorldParams2(context.Background(), params)
			if err != nil {
				return err
			}

			return clientCtx.PrintProto(res)
		},
	}

	flags.AddQueryFlagsToCmd(cmd)

	return cmd
}
